requirements.txt
fastapi>=0.115.0
uvicorn>=0.23.0
python-multipart>=0.0.6
httpx>=0.27.0
redis>=5.0.8
python-dotenv>=1.0.1
androguard>=4.1.1
yara-python>=4.5.1
starlette>=0.38.0
